在蓝牙串口中断中加入野火上位机的数据接收

+++ Seekfree_TC264_Opensource_Library/CODE/protocol.c
@@ -16,11 +16,8 @@
 
 #include "protocol.h"
 #include <string.h>
-//#include ".\motor_control\bsp_motor_control.h"
 #include "Motor.h"
-//#include "./pid/bsp_pid.h"
 #include "PID.h"
-//#include "./tim/bsp_basic_tim.h"
 #include "headfile.h"
 
 
@@ -330,7 +327,7 @@ int8_t receiving_process(void)
       {
         int actual_temp = COMPOUND_32BIT(&frame_data[13]);    // 得到数据
         
-        MotorSetTarget(actual_temp,actual_temp);    // 设置目标值     已做移植替换
+        MotorSetTarget((int16)actual_temp,(int16)actual_temp);    // 设置目标值     已做移植替换
       }
       break;
       
@@ -388,25 +385,25 @@ void set_computer_value(uint8_t cmd, uint8_t ch, void *data, uint8_t num)
   
   sum = check_sum(0, (uint8_t *)&set_packet, sizeof(set_packet));       // 计算包头校验和
   sum = check_sum(sum, (uint8_t *)data, num);                           // 计算参数校验和
-  
-//  uart_putbuff(UART_2, (uint8_t *)&set_packet, sizeof(set_packet));    // 发送数据头  //需要修改
-//  uart_putbuff(UART_2, (uint8_t *)data, num);                          // 发送参数   //需要修改
-//  uart_putbuff(UART_2, (uint8_t *)&sum, sizeof(sum));                  // 发送校验和  //需要修改
 
+  //由于移植过来英飞凌单片机后在发送数据的格式上出现了不太清楚的问题，在这里重写数据发送部分
+//  uart_putbuff(UART_2, (uint8_t *)&set_packet, sizeof(set_packet));    // 发送数据头
+//  uart_putbuff(UART_2, (uint8_t *)data, num);                          // 发送参数
+//  uart_putbuff(UART_2, (uint8_t *)&sum, sizeof(sum));                  // 发送校验和
+
+  //定义并初始化数据
   uint32 my_head=FRAME_HEADER;
   uint8 my_ch=ch;
   uint32 my_len=0x0F;
   uint8 my_cmd =cmd;
-//  uint32 my_data=*data;
   uint8 my_sum=sum;
-
+  //发送数据
   uart_putbuff(UART_2,(uint8 *)&my_head,4);     //包头
   uart_putbuff(UART_2,&my_ch,1);                //通道选择
   uart_putbuff(UART_2,(uint8 *)&my_len,4);      //长度
   uart_putbuff(UART_2,(uint8 *)&my_cmd,1);      //命令
   uart_putbuff(UART_2,(uint8 *)data,4);         //数据
   uart_putbuff(UART_2,(uint8 *)&my_sum,1);      //校验和
-
 }
 
 /**********************************************************************************************/
